+++ src/core/evaluate.py
@@ -6,8 +6,8 @@ from torchvision import transforms
 from PIL import Image
 import os
 import yaml
-import src.core.models as models
-import dataset_preparation
+import src.models as models
+import src.dataset_preparation as dataset_preparation
 import matplotlib.pyplot as plt
 from sklearn.metrics import confusion_matrix, ConfusionMatrixDisplay, precision_score, recall_score, f1_score, roc_auc_score, roc_curve
 from sklearn.preprocessing import label_binarize
