+++ src/core/train.py
@@ -6,8 +6,8 @@ from torch.utils.data import DataLoader, random_split
 from torchvision import transforms
 from config import get_model_config
 from src.core.evaluate import evaluate_model, generate_confusion_matrix
-import src.core.models as models
-import dataset_preparation
+import src.models as models
+import src.dataset_preparation as dataset_preparation
 from tqdm import tqdm
 import yaml
 import numpy as np
