+++ src/hashing/image_generator_using_perceptual_hashing.py
@@ -3,7 +3,7 @@ import numpy as np
 import yaml
 from PIL import Image
 import imagehash
-from dataset_preparation import PerceptualHashingDataset
+from src.dataset_preparation import PerceptualHashingDataset
 import os
